Move rdf to fixtures, lint and typing updates
